--- addons/binding/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/api/IBaseFroniusResponse.java
@@ -1,11 +1,0 @@
-package org.openhab.binding.fronius.internal.api;
-
-/**
- * Basic interface for a response-object from the API
- * 
- * @author rokohl
- *
- */
-public interface IBaseFroniusResponse {
-    public Head getHead();
-}
