+++ addons/binding/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/api/InverterRealtimeResponse.java
@@ -16,23 +16,10 @@ import com.google.gson.annotations.SerializedName;
  *
  * @author Thomas Rokohl - Initial contribution
  */
-public class InverterRealtimeResponse implements IBaseFroniusResponse {
-    @SerializedName("Head")
-    private Head head;
+public class InverterRealtimeResponse extends BaseFroniusResponse {
+
     @SerializedName("Body")
     private InverterRealtimeBody body;
-
-    @Override
-    public Head getHead() {
-        if (head == null) {
-            head = new Head();
-        }
-        return head;
-    }
-
-    public void setHead(Head head) {
-        this.head = head;
-    }
 
     public InverterRealtimeBody getBody() {
         if (body == null) {
